:bug: :art: remove circular dependency in 0.X version
